requirements.txt
jinja2
pandas
numpy
python-dotenv
aiofiles
mysql-connector

